+++ mex/extractors/datenkompass/main.py
@@ -1,3 +1,4 @@
+from collections import deque
 from typing import cast
 
 from dagster import asset
@@ -26,10 +27,6 @@ from mex.extractors.datenkompass.filter import (
     filter_for_organization,
     find_descendant_units,
 )
-from mex.extractors.datenkompass.load import (
-    start_s3_client,
-    write_items_to_xlsx,
-)
 from mex.extractors.datenkompass.models.item import (
     DatenkompassActivity,
     DatenkompassBibliographicResource,
@@ -42,6 +39,7 @@ from mex.extractors.datenkompass.transform import (
 )
 from mex.extractors.pipeline import run_job_in_process
 from mex.extractors.settings import Settings
+from mex.extractors.sinks.s3 import S3XlsxSink
 
 
 @asset(group_name="datenkompass")
@@ -122,11 +120,11 @@ def fetched_merged_activities() -> list[MergedActivity]:
     """Get merged activities."""
     relevant_primary_sources = [
         "blueant",
-        "confluence-vvt",
-        "datscha-web",
-        "ff-projects",
-        "international-projects",
-        "report-server",  # synopse
+        # "confluence-vvt",
+        # "datscha-web",
+        # "ff-projects",
+        # "international-projects",
+        # "report-server",  # synopse
     ]
     entity_type = ["MergedActivity"]
     primary_source_ids = get_relevant_primary_source_ids(relevant_primary_sources)
@@ -172,7 +170,7 @@ def fetched_merged_resources_by_primary_source(
     filtered_merged_organizational_unit_ids: list[str],
 ) -> dict[str, list[MergedResource]]:
     """Get merged resources as dictionary."""
-    relevant_primary_sources = ["open-data", "report-server"]
+    relevant_primary_sources = ["open-data"]  # , "report-server"]
     entity_type = ["MergedResource"]
     merged_resources_by_primary_source: dict[str, list[MergedResource]] = {}
     for rps in relevant_primary_sources:
@@ -260,21 +258,25 @@ def transform_resources_to_datenkompass_resources(
 
 
 @asset(group_name="datenkompass")
-def load_activities(
+def publish_to_s3_xlsx(
     transform_activities_to_datenkompass_activities: list[DatenkompassActivity],
     transform_bibliographic_resources_to_datenkompass_bibliographic_resources: list[
         DatenkompassBibliographicResource
     ],
     transform_resources_to_datenkompass_resources: list[DatenkompassResource],
 ) -> None:
-    """Write items to S3."""
-    s3_client = start_s3_client()
-    write_items_to_xlsx(transform_activities_to_datenkompass_activities, s3_client)
-    write_items_to_xlsx(
-        transform_bibliographic_resources_to_datenkompass_bibliographic_resources,
-        s3_client,
+    """Write items to S3 xlsx."""
+    settings = Settings.get()
+    sep = settings.datenkompass.list_delimiter
+    s3xlsx = S3XlsxSink(separator=sep, sort_columns_alphabetically=False)
+    deque(s3xlsx.load(transform_activities_to_datenkompass_activities), maxlen=0)
+    deque(
+        s3xlsx.load(
+            transform_bibliographic_resources_to_datenkompass_bibliographic_resources,
+        ),
+        maxlen=0,
     )
-    write_items_to_xlsx(transform_resources_to_datenkompass_resources, s3_client)
+    deque(s3xlsx.load(transform_resources_to_datenkompass_resources), maxlen=0)
 
 
 @entrypoint(Settings)