+++ mex/extractors/datenkompass/transform.py
@@ -290,7 +290,6 @@ def transform_activities(
                 ),
                 format="Sonstiges",
                 identifier=item.identifier,
-                entityType=item.entityType,
             ),
         )
     return datenkompass_activities
@@ -365,7 +364,6 @@ def transform_bibliographic_resources(
                     "voraussichtlich Ende 2025 verfügbar sein."
                 ),
                 identifier=item.identifier,
-                entityType=item.entityType,
             ),
         )
     return datenkompass_bibliographic_recources
@@ -461,7 +459,6 @@ def transform_resources(
                         "voraussichtlich Ende 2025 verfügbar sein."
                     ),
                     identifier=item.identifier,
-                    entityType=item.entityType,
                 ),
             )
     return datenkompass_recources